requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
bzip2=1.0.8=h1de35cc_0
ca-certificates=2020.1.1=0
cairo=1.16.0=h0ab9d94_1001
certifi=2020.4.5.1=py38_0
cycler=0.10.0=py38_0
cython=0.29.17=py38h0a44026_0
ffmpeg=4.2.2=h97e5cf8_0
fontconfig=2.13.1=h1027ab8_1000
freetype=2.9.1=hb4e5f40_0
gettext=0.19.8.1=h15daf44_3
giflib=5.1.4=h1de35cc_1
glib=2.63.1=hd977a24_0
gmp=6.1.2=hb37e062_1
gnutls=3.6.5=h91ad68e_1002
graphite2=1.3.13=h2098e52_0
harfbuzz=2.4.0=h92b87b8_1
hdf5=1.10.6=nompi_h3e39495_100
icu=58.2=h0a44026_3
intel-openmp=2020.1=216
jasper=1.900.1=h1f36771_4
jpeg=9d=h0b31af3_0
kiwisolver=1.2.0=py38h04f5b5a_0
lame=3.100=h1de35cc_0
libblas=3.8.0=16_openblas
libcblas=3.8.0=16_openblas
libcxx=10.0.0=1
libedit=3.1.20181209=hb402a30_0
libffi=3.3=h0a44026_1
libgfortran=4.0.0=2
libiconv=1.15=hdd342a3_7
liblapack=3.8.0=16_openblas
liblapacke=3.8.0=16_openblas
libopenblas=0.3.9=h3d69b6c_0
libopencv=4.2.0=py38_6
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_1
libvpx=1.7.0=h378b8a2_0
libwebp=1.0.2=h801f6e5_1
libxml2=2.9.9=hf6e021a_1
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.2=h0a44026_0
matplotlib-base=3.1.3=py38h9aa3819_0
mkl=2020.1=216
ncurses=6.2=h0a44026_1
nettle=3.4.1=h3018a27_0
ninja=1.9.0=py38h04f5b5a_0
numpy=1.18.5=py38h1f821a2_0
olefile=0.46=py_0
opencv=4.2.0=py38_6
openh264=2.1.0=hd9629dc_0
openssl=1.1.1g=h1de35cc_0
pbkdf2=1.3=py38_1
pcre=8.43=h0a44026_0
pillow=7.1.2=py38h4655f20_0
pip=20.0.2=py38_3
pixman=0.38.0=h1de35cc_0
py-opencv=4.2.0=py38h23f93f0_6
pycocotools=2.0.0=py38h0b31af3_1001
pycrypto=2.6.1=py38h64e0658_1004
pyparsing=2.4.7=py_0
python=3.8.1=h359304d_1
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.4.0=py3.8_0
readline=7.0=h1de35cc_5
setuptools=47.1.1=py38_0
six=1.15.0=py_0
sqlite=3.31.1=h5c1f38d_1
tk=8.6.8=ha441bb4_0
torchvision=0.5.0=py38_cpu
tornado=6.0.4=py38h1de35cc_1
tqdm=4.42.1=py_0
wheel=0.34.2=py38_0
x264=1!157.20191217=h1de35cc_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.4=h1990bb4_3
